jwt_validation/requirements.txt
async-timeout==4.0.2
bottle==0.12.19
certifi==2021.10.8
charset-normalizer==2.0.12
Deprecated==1.2.13
idna==3.3
packaging==21.3
Paste==3.5.0
pydantic==1.9.0
PyJWT==2.3.0
pyparsing==3.0.8
python-http-client==3.3.7
PyYAML==6.0
redis==4.3.0
requests==2.27.1
sendgrid==6.9.7
six==1.16.0
starkbank-ecdsa==2.0.3
typing_extensions==4.2.0
urllib3==1.26.9
wrapt==1.14.1
xmltodict==0.13.0
